+++ tests/TestCase/Controller/Component/UploaderComponentTest.php
@@ -24,14 +24,10 @@ use Tools\Filesystem;
 
 /**
  * UploaderComponentTest class
+ * @property \MeTools\Controller\Component\UploaderComponent $Component
  */
 class UploaderComponentTest extends ComponentTestCase
 {
-    /**
-     * @var \MeTools\Controller\Component\UploaderComponent
-     */
-    protected $Component;
-
     /**
      * Internal method to create a file and get a `UploadedFile` instance
      * @param int $error Error for this file
@@ -79,9 +75,7 @@ class UploaderComponentTest extends ComponentTestCase
     public function testFindTargetFilename(): void
     {
         $Filesystem = new Filesystem();
-        $findTargetFilenameMethod = function (string $filename): string {
-            return $this->invokeMethod($this->Component, 'findTargetFilename', [$filename]);
-        };
+        $findTargetFilenameMethod = fn(string $filename): string => $this->invokeMethod($this->Component, 'findTargetFilename', [$filename]);
 
         $file1 = UPLOADS . 'target.txt';
         $file2 = UPLOADS . 'target_1.txt';
@@ -105,23 +99,6 @@ class UploaderComponentTest extends ComponentTestCase
         //Creates the first file
         $Filesystem->createFile($file1);
         $this->assertEquals($file2, $findTargetFilenameMethod($file1));
-    }
-
-    /**
-     * Tests for `set()` method
-     * @test
-     */
-    public function testSet(): void
-    {
-        $file = $this->createFile();
-
-        $current = error_reporting(E_ALL & ~E_USER_DEPRECATED);
-        $this->Component->set($file);
-        error_reporting($current);
-
-        $this->expectDeprecation();
-        $this->expectExceptionMessage('Deprecated. Use instead `setFile()');
-        $this->Component->set($file);
     }
 
     /**
@@ -220,8 +197,7 @@ class UploaderComponentTest extends ComponentTestCase
             ->setMethods(['moveTo'])
             ->getMock();
 
-        $UploadedFile->method('moveTo')
-            ->willThrowException(new UploadedFileErrorException());
+        $UploadedFile->method('moveTo')->willThrowException(new UploadedFileErrorException());
 
         $this->assertFalse($this->Component->setFile($UploadedFile)->save(UPLOADS));
         $this->assertSame('The file was not successfully moved to the target directory', $this->Component->getError());